+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.2.0] - 2021-06-17
+### Added
+ - Example Ruby scripts.
+ - Interactive Clojure interpreter.
+
+
 ## [0.1.0] - 2021-03-08
 ### Added
  - Interactive Ruby interpreter.